+++ tests/testthat/test-expectations.r
@@ -9,12 +9,17 @@ test_that("errors are caught with throws_error", {
 
   res <- throws_error("Yes")(stop("No"))
   expect_that(res$passed, is_false())
+
+  res <- throws_error(NA)(stop())
+  expect_that(res$passed, is_false())
+  expect_match(res$failure_msg, "expected no errors:")
+  expect_match(res$success_msg, "no errors raised")
 })
 
 test_that("failure to throw an error is a failure", {
   res <- throws_error()(log(1))
   expect_that(res$passed, is_false())
-  expect_output(res, "code didn't raise an error")
+  expect_match(res$failure_msg, "no errors raised")
 
   res <- throws_error("error")(log(1))
   expect_that(res$passed, is_false())