+++ R/reporter-summary.r
@@ -23,13 +23,14 @@ NULL
 SummaryReporter <- setRefClass("SummaryReporter", contains = "Reporter",
   fields = list(
     "failures" = "list",
+    "skips" = "list",
     "n" = "integer",
     "has_tests" = "logical",
     "max_reports" = "numeric",
     "show_praise" = "logical"),
 
   methods = list(
-    initialize = function(max_reports = Inf, ...) {
+    initialize = function(max_reports = getOption("testthat.summary.max_reports", 15L), ...) {
       max_reports <<- max_reports
       show_praise <<- TRUE
       callSuper(...)
@@ -45,6 +46,7 @@ SummaryReporter <- setRefClass("SummaryReporter", contains = "Reporter",
 
     start_reporter = function() {
       failures <<- list()
+      skips <<- list()
       has_tests <<- FALSE
       n <<- 0L
     },
@@ -53,6 +55,8 @@ SummaryReporter <- setRefClass("SummaryReporter", contains = "Reporter",
       callSuper(result)
       has_tests <<- TRUE
       if (result$skipped) {
+        result$test <- if (is.null(test)) "(unknown)" else test
+        skips <<- c(skips, list(result))
         cat(colourise("S", "skipped"))
         return()
       }
@@ -75,6 +79,12 @@ SummaryReporter <- setRefClass("SummaryReporter", contains = "Reporter",
       if (n == 0) {
         if (!has_tests)
           return()
+
+        if (length(skips) > 0L) {
+          cat(colourise("\nSkip:", "skipped"), "\n\n")
+          cat_reports(skips, skip_summary, "\n")
+        }
+
         cat("\n")
         if (show_praise && runif(1) < 0.1) {
           cat(colourise(praise(), "passed"), "\n")
@@ -83,11 +93,7 @@ SummaryReporter <- setRefClass("SummaryReporter", contains = "Reporter",
         }
       } else {
         cat("\n")
-
-        reports <- vapply(seq_len(n), function(i) {
-          failure_summary(failures[[i]], labels[i])
-        }, character(1))
-        cat(paste(reports, collapse = "\n\n"), "\n", sep = "")
+        cat_reports(failures, failure_summary, "\n\n")
 
         if (show_praise && runif(1) < 0.25) {
           cat("\n", colourise(encourage(), "error"), "\n", sep = "")
@@ -98,3 +104,17 @@ SummaryReporter <- setRefClass("SummaryReporter", contains = "Reporter",
 )
 
 labels <- c(1:9, letters, LETTERS)
+
+cat_reports <- function(reports, summary_fun, collapse) {
+  report_summary <- vapply(
+    seq_len(min(length(reports), length(labels))),
+    function(i) {
+      summary_fun(reports[[i]], labels[i])
+    }, character(1L))
+  cat(paste(report_summary, collapse = collapse), "\n", sep = "")
+
+  extra_reports <- length(reports) - length(labels)
+  if (extra_reports > 0L) {
+    cat("  ... and ", extra_reports, " more\n", sep = "")
+  }
+}