+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * `set_reporter()` returns old reporter invisibly (#358, @krlmlr).
 
+* New option `testthat.summary.max_reports` that limits the number of reports printed by the summary reporter, default: 15 (@krlmlr, #354).
+
 * Fix failure message for `throws_error` in case where no error is raised (#342, @nealrichardson).
 
 * Added [Catch](https://github.com/philsquared/Catch) for unit testing of C++ code.