+++ NEWS.md
@@ -2,6 +2,8 @@
 
 * Fixed minor issues with `throws_error(NA)` and `expect_error(..., NA)` (@krlmlr, #317).
 
+* New option `testthat.summary.max_reports` that limits the number of reports printed by the summary reporter, default: 15 (@krlmlr, #354).
+
 * Fix failure message for `throws_error` in case where no error is raised (@nealrichardson, #342).
 
 * Added [Catch](https://github.com/philsquared/Catch) for unit testing of C++ code.