+++ NEWS.md
@@ -7,6 +7,10 @@
 * New `FailReporter` to stop in case of failures or errors after all tests
   (#308, @krlmlr).
 
+* Comparing integers to non-numbers doesn't raise errors anymore, and falls back to
+  string comparison if objects have different lengths. Complex numbers are compared
+  using the same routine (#309, @krlmlr).
+
 * New option `testthat.summary.max_reports` that limits the number of reports printed by the summary reporter, default: 15 (@krlmlr, #354).
 
 * Fix failure message for `throws_error` in case where no error is raised (#342, @nealrichardson).